+++ cia-core/src/main/java/top/ourisland/creepersiarena/core/command/CiaCommandRoots.java
@@ -3,8 +3,9 @@ package top.ourisland.creepersiarena.core.command;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import io.papermc.paper.command.brigadier.CommandSourceStack;
 import top.ourisland.creepersiarena.core.bootstrap.BootstrapRuntime;
-import top.ourisland.creepersiarena.core.command.handler.AdminCommandHandlers;
-import top.ourisland.creepersiarena.core.command.handler.PlayerCommandHandlers;
+import top.ourisland.creepersiarena.core.command.handler.AdminHandlers;
+import top.ourisland.creepersiarena.core.command.handler.CommandHandlerContext;
+import top.ourisland.creepersiarena.core.command.handler.PlayerHandlers;
 import top.ourisland.creepersiarena.core.command.tree.AdminCommandTree;
 import top.ourisland.creepersiarena.core.command.tree.PlayerCommandTree;
 
@@ -17,8 +18,10 @@ public final class CiaCommandRoots {
     private final AdminCommandTree adminTree;
 
     public CiaCommandRoots(BootstrapRuntime rt) {
-        var playerHandlers = new PlayerCommandHandlers(rt);
-        var adminHandlers = new AdminCommandHandlers(rt);
+        var context = new CommandHandlerContext(rt);
+        var playerHandlers = new PlayerHandlers(context);
+        var adminHandlers = new AdminHandlers(context);
+
         this.adminTree = new AdminCommandTree(rt, adminHandlers);
         this.playerTree = new PlayerCommandTree(rt, playerHandlers);
     }
