+++ cia-core/src/main/java/top/ourisland/creepersiarena/core/command/tree/AbilityAdminCommandTree.java
@@ -5,7 +5,7 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import top.ourisland.creepersiarena.core.bootstrap.BootstrapRuntime;
 import top.ourisland.creepersiarena.core.command.argument.CiaArguments;
-import top.ourisland.creepersiarena.core.command.handler.AdminCommandHandlers;
+import top.ourisland.creepersiarena.core.command.handler.admin.AbilityAdminHandlers;
 import top.ourisland.creepersiarena.core.command.permission.CiaPermissions;
 import top.ourisland.creepersiarena.core.command.suggestion.RegistrySuggestions;
 
@@ -15,32 +15,32 @@ import top.ourisland.creepersiarena.core.command.suggestion.RegistrySuggestions;
 public final class AbilityAdminCommandTree {
 
     private final BootstrapRuntime rt;
-    private final AdminCommandHandlers admin;
+    private final AbilityAdminHandlers ability;
 
     public AbilityAdminCommandTree(
             BootstrapRuntime rt,
-            AdminCommandHandlers admin
+            AbilityAdminHandlers ability
     ) {
         this.rt = rt;
-        this.admin = admin;
+        this.ability = ability;
     }
 
     public LiteralArgumentBuilder<CommandSourceStack> build(String literal) {
         return Commands.literal(literal)
                 .requires(source -> CiaArguments.hasPermission(source, CiaPermissions.ADMIN_ABILITY))
                 .executes(ctx -> {
-                    admin.abilityList(CiaArguments.sender(ctx));
+                    ability.abilityList(CiaArguments.sender(ctx));
                     return 1;
                 })
                 .then(Commands.literal("list")
                         .executes(ctx -> {
-                            admin.abilityList(CiaArguments.sender(ctx));
+                            ability.abilityList(CiaArguments.sender(ctx));
                             return 1;
                         })
                 )
                 .then(Commands.literal("reload")
                         .executes(ctx -> {
-                            admin.abilityReload(CiaArguments.sender(ctx));
+                            ability.abilityReload(CiaArguments.sender(ctx));
                             return 1;
                         })
                 )
@@ -54,12 +54,12 @@ public final class AbilityAdminCommandTree {
                 .then(CiaArguments.ciaKey("ability_id")
                         .suggests((_, builder) -> RegistrySuggestions.abilityIds(rt, builder))
                         .executes(ctx -> {
-                            admin.abilityInfo(CiaArguments.sender(ctx), CiaArguments.abilityId(ctx, "ability_id"));
+                            ability.abilityInfo(CiaArguments.sender(ctx), CiaArguments.abilityId(ctx, "ability_id"));
                             return 1;
                         })
                 )
                 .executes(ctx -> {
-                    admin.abilityUsage(CiaArguments.sender(ctx));
+                    ability.abilityUsage(CiaArguments.sender(ctx));
                     return 1;
                 });
     }
@@ -70,15 +70,15 @@ public final class AbilityAdminCommandTree {
                         .suggests((_, builder) -> RegistrySuggestions.abilityIds(rt, builder))
                         .executes(ctx -> {
                             if (enabled) {
-                                admin.enableAbility(CiaArguments.sender(ctx), CiaArguments.abilityId(ctx, "ability_id"));
+                                ability.enableAbility(CiaArguments.sender(ctx), CiaArguments.abilityId(ctx, "ability_id"));
                             } else {
-                                admin.disableAbility(CiaArguments.sender(ctx), CiaArguments.abilityId(ctx, "ability_id"));
+                                ability.disableAbility(CiaArguments.sender(ctx), CiaArguments.abilityId(ctx, "ability_id"));
                             }
                             return 1;
                         })
                 )
                 .executes(ctx -> {
-                    admin.abilityUsage(CiaArguments.sender(ctx));
+                    ability.abilityUsage(CiaArguments.sender(ctx));
                     return 1;
                 });
     }
