+++ cia-core/src/main/java/top/ourisland/creepersiarena/core/command/tree/ExtensionAdminCommandTree.java
@@ -6,7 +6,7 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import top.ourisland.creepersiarena.core.bootstrap.BootstrapRuntime;
 import top.ourisland.creepersiarena.core.command.argument.CiaArguments;
-import top.ourisland.creepersiarena.core.command.handler.AdminCommandHandlers;
+import top.ourisland.creepersiarena.core.command.handler.admin.ExtensionAdminHandlers;
 import top.ourisland.creepersiarena.core.command.permission.CiaPermissions;
 import top.ourisland.creepersiarena.core.command.suggestion.RegistrySuggestions;
 
@@ -16,26 +16,26 @@ import top.ourisland.creepersiarena.core.command.suggestion.RegistrySuggestions;
 public final class ExtensionAdminCommandTree {
 
     private final BootstrapRuntime rt;
-    private final AdminCommandHandlers admin;
+    private final ExtensionAdminHandlers extension;
 
     public ExtensionAdminCommandTree(
             BootstrapRuntime rt,
-            AdminCommandHandlers admin
+            ExtensionAdminHandlers extension
     ) {
         this.rt = rt;
-        this.admin = admin;
+        this.extension = extension;
     }
 
     public LiteralArgumentBuilder<CommandSourceStack> build(String literal) {
         return Commands.literal(literal)
                 .requires(source -> CiaArguments.hasPermission(source, CiaPermissions.ADMIN_EXTENSION))
                 .executes(ctx -> {
-                    admin.extensionsList(CiaArguments.sender(ctx));
+                    extension.extensionsList(CiaArguments.sender(ctx));
                     return 1;
                 })
                 .then(Commands.literal("list")
                         .executes(ctx -> {
-                            admin.extensionsList(CiaArguments.sender(ctx));
+                            extension.extensionsList(CiaArguments.sender(ctx));
                             return 1;
                         })
                 )
@@ -43,18 +43,18 @@ public final class ExtensionAdminCommandTree {
                         .then(CiaArguments.word("extension_id")
                                 .suggests((_, builder) -> RegistrySuggestions.extensionIds(rt, builder))
                                 .executes(ctx -> {
-                                    admin.extensionInfo(CiaArguments.sender(ctx), StringArgumentType.getString(ctx, "extension_id"));
+                                    extension.extensionInfo(CiaArguments.sender(ctx), StringArgumentType.getString(ctx, "extension_id"));
                                     return 1;
                                 })
                         )
                         .executes(ctx -> {
-                            admin.extensionInfo(CiaArguments.sender(ctx), "");
+                            extension.extensionInfo(CiaArguments.sender(ctx), "");
                             return 1;
                         })
                 )
                 .then(Commands.literal("dump")
                         .executes(ctx -> {
-                            admin.extensionsDump(CiaArguments.sender(ctx));
+                            extension.extensionsDump(CiaArguments.sender(ctx));
                             return 1;
                         })
                 );
