+++ cia-core/src/main/java/top/ourisland/creepersiarena/core/command/tree/GameAdminCommandTree.java
@@ -9,7 +9,7 @@ import io.papermc.paper.command.brigadier.Commands;
 import top.ourisland.creepersiarena.core.bootstrap.BootstrapRuntime;
 import top.ourisland.creepersiarena.core.command.CiaCommandConstants;
 import top.ourisland.creepersiarena.core.command.argument.CiaArguments;
-import top.ourisland.creepersiarena.core.command.handler.AdminCommandHandlers;
+import top.ourisland.creepersiarena.core.command.handler.admin.GameAdminHandlers;
 import top.ourisland.creepersiarena.core.command.permission.CiaPermissions;
 import top.ourisland.creepersiarena.core.command.suggestion.CiaSuggestions;
 import top.ourisland.creepersiarena.core.command.suggestion.RegistrySuggestions;
@@ -20,21 +20,21 @@ import top.ourisland.creepersiarena.core.command.suggestion.RegistrySuggestions;
 public final class GameAdminCommandTree {
 
     private final BootstrapRuntime rt;
-    private final AdminCommandHandlers admin;
+    private final GameAdminHandlers game;
 
     public GameAdminCommandTree(
             BootstrapRuntime rt,
-            AdminCommandHandlers admin
+            GameAdminHandlers game
     ) {
         this.rt = rt;
-        this.admin = admin;
+        this.game = game;
     }
 
     public LiteralArgumentBuilder<CommandSourceStack> build(String literal) {
         return Commands.literal(literal)
                 .requires(source -> CiaArguments.hasPermission(source, CiaPermissions.ADMIN_GAME))
                 .executes(ctx -> {
-                    admin.gameUsage(CiaArguments.sender(ctx));
+                    game.gameUsage(CiaArguments.sender(ctx));
                     return 1;
                 })
                 .then(mode())
@@ -52,12 +52,12 @@ public final class GameAdminCommandTree {
                 .then(CiaArguments.ciaKey("mode_id")
                         .suggests((_, builder) -> RegistrySuggestions.modeIds(rt, builder))
                         .executes(ctx -> {
-                            admin.mode(CiaArguments.sender(ctx), CiaArguments.modeId(ctx, "mode_id"));
+                            game.mode(CiaArguments.sender(ctx), CiaArguments.modeId(ctx, "mode_id"));
                             return 1;
                         })
                 )
                 .executes(ctx -> {
-                    admin.modeUsage(CiaArguments.sender(ctx));
+                    game.modeUsage(CiaArguments.sender(ctx));
                     return 1;
                 });
     }
@@ -69,15 +69,15 @@ public final class GameAdminCommandTree {
                         .suggests((_, builder) -> RegistrySuggestions.arenaIds(rt, builder))
                         .executes(ctx -> {
                             try {
-                                admin.arena(CiaArguments.sender(ctx), CiaArguments.arenaId(ctx, "arena_id"));
+                                game.arena(CiaArguments.sender(ctx), CiaArguments.arenaId(ctx, "arena_id"));
                             } catch (IllegalArgumentException exception) {
-                                admin.invalidArena(CiaArguments.sender(ctx), exception.getMessage());
+                                game.invalidArena(CiaArguments.sender(ctx), exception.getMessage());
                             }
                             return 1;
                         })
                 )
                 .executes(ctx -> {
-                    admin.arenaUsage(CiaArguments.sender(ctx));
+                    game.arenaUsage(CiaArguments.sender(ctx));
                     return 1;
                 });
     }
@@ -89,15 +89,15 @@ public final class GameAdminCommandTree {
                         .suggests((_, builder) -> RegistrySuggestions.arenaIds(rt, builder))
                         .executes(ctx -> {
                             try {
-                                admin.skip(CiaArguments.sender(ctx), CiaArguments.arenaId(ctx, "arena_id"));
+                                game.skip(CiaArguments.sender(ctx), CiaArguments.arenaId(ctx, "arena_id"));
                             } catch (IllegalArgumentException exception) {
-                                admin.invalidArena(CiaArguments.sender(ctx), exception.getMessage());
+                                game.invalidArena(CiaArguments.sender(ctx), exception.getMessage());
                             }
                             return 1;
                         })
                 )
                 .executes(ctx -> {
-                    admin.skip(CiaArguments.sender(ctx), null);
+                    game.skip(CiaArguments.sender(ctx), null);
                     return 1;
                 });
     }
@@ -107,12 +107,12 @@ public final class GameAdminCommandTree {
                 .requires(source -> CiaArguments.hasPermission(source, CiaPermissions.ADMIN_COOLDOWN))
                 .then(RequiredArgumentBuilder.<CommandSourceStack, Double>argument("factor", DoubleArgumentType.doubleArg(0.0D))
                         .executes(ctx -> {
-                            admin.setCooldownFactor(CiaArguments.sender(ctx), DoubleArgumentType.getDouble(ctx, "factor"));
+                            game.setCooldownFactor(CiaArguments.sender(ctx), DoubleArgumentType.getDouble(ctx, "factor"));
                             return 1;
                         })
                 )
                 .executes(ctx -> {
-                    admin.cooldownUsage(CiaArguments.sender(ctx));
+                    game.cooldownUsage(CiaArguments.sender(ctx));
                     return 1;
                 });
     }
@@ -122,12 +122,12 @@ public final class GameAdminCommandTree {
                 .requires(source -> CiaArguments.hasPermission(source, CiaPermissions.ADMIN_REGENERATION))
                 .then(RequiredArgumentBuilder.<CommandSourceStack, Double>argument("factor", DoubleArgumentType.doubleArg(0.0D))
                         .executes(ctx -> {
-                            admin.setRegenerationFactor(CiaArguments.sender(ctx), DoubleArgumentType.getDouble(ctx, "factor"));
+                            game.setRegenerationFactor(CiaArguments.sender(ctx), DoubleArgumentType.getDouble(ctx, "factor"));
                             return 1;
                         })
                 )
                 .executes(ctx -> {
-                    admin.regenerationStatus(CiaArguments.sender(ctx));
+                    game.regenerationStatus(CiaArguments.sender(ctx));
                     return 1;
                 });
     }
@@ -137,19 +137,19 @@ public final class GameAdminCommandTree {
                 .requires(source -> CiaArguments.hasPermission(source, CiaPermissions.ADMIN_MUTATION))
                 .then(Commands.literal("trigger")
                         .executes(ctx -> {
-                            admin.triggerMutation(CiaArguments.sender(ctx));
+                            game.triggerMutation(CiaArguments.sender(ctx));
                             return 1;
                         })
                 )
                 .then(RequiredArgumentBuilder.<CommandSourceStack, Boolean>argument("enabled", BoolArgumentType.bool())
                         .suggests((_, builder) -> CiaSuggestions.staticValues(builder, CiaCommandConstants.BOOLEAN_SUGGESTIONS))
                         .executes(ctx -> {
-                            admin.setMutationEnabled(CiaArguments.sender(ctx), BoolArgumentType.getBool(ctx, "enabled"));
+                            game.setMutationEnabled(CiaArguments.sender(ctx), BoolArgumentType.getBool(ctx, "enabled"));
                             return 1;
                         })
                 )
                 .executes(ctx -> {
-                    admin.mutationStatus(CiaArguments.sender(ctx));
+                    game.mutationStatus(CiaArguments.sender(ctx));
                     return 1;
                 });
     }
@@ -160,12 +160,12 @@ public final class GameAdminCommandTree {
                 .then(RequiredArgumentBuilder.<CommandSourceStack, Boolean>argument("enabled", BoolArgumentType.bool())
                         .suggests((_, builder) -> CiaSuggestions.staticValues(builder, CiaCommandConstants.BOOLEAN_SUGGESTIONS))
                         .executes(ctx -> {
-                            admin.entrance(CiaArguments.sender(ctx), BoolArgumentType.getBool(ctx, "enabled"));
+                            game.entrance(CiaArguments.sender(ctx), BoolArgumentType.getBool(ctx, "enabled"));
                             return 1;
                         })
                 )
                 .executes(ctx -> {
-                    admin.entranceUsage(CiaArguments.sender(ctx));
+                    game.entranceUsage(CiaArguments.sender(ctx));
                     return 1;
                 });
     }
